third_party/blink/web_tests/http/tests/local/formdata/send-form-data-with-string-containing-null-expected.txt

Test that we can send textarea data containing null characters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS the string containing a null byte is echoed correctly.
PASS successfullyParsed is true

TEST COMPLETE
